ADD highlight selected on preview
NOTE this is a one-way direction, from properties sidebar to preview.
In order to make it both-ways, you need to make PropertiesSidebar accept
selected props as well.

NOTE currently Preview renders nodes with `{@html source}`, so it is
difficult to add behaviors. We should make the code be more like svelte.